# Optimizations: low-hanging fruits