--- a/pyproject.toml
+++ b/pyproject.toml
@@ -91,8 +91,7 @@ dev = [
   "types-python-dateutil",
   "types-pytz",
   "types-PyYAML",
-  "types-requests",
-  "types-orjson"
+  "types-requests"
 ]
 
 [project.urls]