@skshetry 2024-07-25
orjson serializes to bytes, so it's not quite a drop-in replacement.