@dreadatour 2024-07-25
> orjson serializes to bytes, so it's not quite a drop-in replacement.

Hm, yes, I remember it now. This was one of the reasons why we do choose `ujson` before. 🤔 Checking.