@skshetry 2024-07-25
Moving off of ujson is a right thing to do, as it is now in _maintenance_ mode.